Redirection and improved logic
