+++ .gitignore
@@ -1,2 +1,3 @@
 .env
 test.py
+**/venv
