+++ backend/.env.example
@@ -2,7 +2,6 @@
 This is a template for the .env file.
 Please make a copy of this file and rename it to .env
 ```
-
 ```
 GOOGLE_CLIENT_ID & GOOGLE_CLIENT_SECRET are obtained from Google API Console. 
 You can follow the instructions here to obtain them: https://developers.google.com/identity/sign-in/web/sign-in
