+++ manifest.json
@@ -3,17 +3,18 @@
 	"name": "PhishMeNot",
 	"version": "0.1.0",
 	"description": "PhishMeNot allows you to surf the web and check your email safely by warning you of suspected phishing attempts.",
-	"permissions": ["storage", "activeTab", "webRequest", "identity", "cookies", "tabs", "notifications", "unlimitedStorage"],
+	"permissions": ["storage", "activeTab", "webNavigation", "webRequest", "identity", "cookies", "tabs", "notifications", "unlimitedStorage"],
 	"host_permissions": ["<all_urls>"],
 	"background": {
-	  "service_worker": "static/js/background.js"
+	  "service_worker": "background.js"
 	},
 	"action": {
 	  "default_popup": "views/auth/auth.html",
 	  "default_icon": {
-		"16": "static/images/icon_16.png",
-		"32": "static/images/icon_32.png",
-		"192": "static/images/icon_192.png"
+		"16": "static/images/logo/icon_16.png",
+		"32": "static/images/logo/icon_32.png",
+		"48": "static/images/logo/icon_48.png",
+		"128": "static/images/logo/icon_128.png"
 	  }
 	}
   }
