--- popup.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>PhishMeNot</title>
-  <link rel="stylesheet" href="static/css/main.css">
-</head>
-<body>
-  <div class="app-root">
-    
-  </div>
-  <script src="static/js/popup.js" defer="defer" charset="UTF-8"></script>
-</body>
-</html>
