--- static/js/popup.js
@@ -1,9 +1,0 @@
-// FishMeNot Login Button
-document.getElementsByClassName('fishmenot-sign-in-button')[0].addEventListener('click', function() {
-    chrome.runtime.sendMessage({action: "phishmenotLogin"});
-  });
-  
-  // Google Login button
-  document.getElementsByClassName('g-sign-in-button')[0].addEventListener('click', function() {
-    chrome.runtime.sendMessage({action: "googleLogin"});
-  });
