+++ views/auth/auth.css
@@ -1,3 +1,7 @@
+.logo {
+  text-align: center;
+}
+
 .sign-in-text{
   margin-top: 1rem;
   color: var(--ivory);
@@ -8,7 +12,7 @@
 }
 
 .sign-in-btns{
-  margin-top: 2rem;
+  margin-top: 1rem;
   display: block;
   justify-content: center;
   width: 100%;
@@ -115,17 +119,4 @@
   display: inline-block;
   text-align: center;
   width: 180px;
-}
-
-/**FOOTER**/
-
-footer {
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  height: 50px;
-  color: var(--ivory);
-  text-align: center;
-  font-size: 0.8rem;
-  padding-top: 1rem;
 }
