+++ views/auth/auth.html
@@ -12,7 +12,9 @@
 <body>
   <div class="card">
     <header>
-      <img src="../../static/images/icon_192.png" alt="PhishMeNot Logo" style="width: 100px;">
+      <div class="logo">
+        <img src="../../static/images/logo/icon_128.png" alt="PhishMeNot Logo">
+      </div>
     </header>
     <div class="body">
       <div class="sign-in-wrapper">
@@ -25,7 +27,7 @@
             <div class="fishmenot-sign-in-button">
               <div class=content-wrapper>
                 <div class='logo-wrapper'>
-                  <img src='../../static/images/icon_32.png'>
+                  <img src='../../static/images/logo/icon_32.png'>
                 </div>
                 <span class='text-container'>
                   <span>Sign in with PhishMeNot</span>
@@ -52,7 +54,7 @@
       <p>PhisMeNot 2023 All Right reserved</p>
     </footer>
   </div>
-  <script src="../../static/js/popup.js" defer="defer" charset="UTF-8"></script>
+  <script src="auth.js" defer="defer" charset="UTF-8"></script>
 </body>
 
 </html>
