+++ views/dashboard/dashboard.html
@@ -3,57 +3,59 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Authentication</title>
-    <link rel="stylesheet" href="style.css" />
+    <title>Dashboard</title>
+    <link rel="stylesheet" href="../../static/css/main.css">
+    <link rel="stylesheet" href="dashboard.css" />
   </head>
   <body>
     <div class="card">
       <header>
-        <h3>PhishMeNot/Auth</h3>
+        <div class="logo">
+          <img src="../../static/images/logo/icon_128.png" alt="PhishMeNot Logo">
+        </div>
       </header>
-      <div class="body">
+      <div class="content">
         <div class="btn_filter">
           <button>Filter</button>
         </div>
         <div class="list_history">
-          
           <div class="item">
-            <img src="../assets/ok.png" alt="" />
+            <img src="../../static/images/ok.png" alt="" />
             <div class="info">
-              <h4>https://test.com</h4>
-              <p>clean</p>
+              <h4>http://test.com</h4>
+              <p>Phishing</p>
             </div>
             <div class="icons">
               <button>
-                <img src="../assets/ok.png" alt="" />
+                <img src="../../static/images/ok.png" />
               </button>
-              <h3>75%</h3>
+              <h3 class="percentage">75%</h3>
             </div>
           </div>
 
           <div class="item">
-            <img src="../assets/ok.png" alt="" />
+            <img src="../../static/images/ok.png" alt="" />
             <div class="info">
               <h4>https://test.com</h4>
               <p>clean</p>
             </div>
             <div class="icons">
               <button>
-                <img src="../assets/ok.png" alt="" />
+                <img src="../../static/images/ok.png" alt="" />
               </button>
-              <h3>75%</h3>
+              <h3 class="percentage">15%</h3>
             </div>
           </div>
 
         </div>
-        <div class="static">
+        <div class="statistics">
           <div class="item">
             <span class="circle ok">
               <span class="circle_in">
                 20
               </span>
             </span>
-            <p>Site non dangereux</p>
+            <p class="danger-tag">Safe websites</p>
           </div>
           <div class="item">
             <span class="circle malicious">
@@ -61,7 +63,7 @@
                 5
               </span>
             </span>
-            <p>Site dangereux</p>
+            <p class="danger-tag">Dangerous websites</p>
           </div>
           <div class="item">
             <span class="circle white_list">
@@ -69,14 +71,13 @@
                 10
               </span>
             </span>
-            <p>Ma white-list</p>
+            <p class="whitelist">My whitelist</p>
           </div>
         </div>
       </div>
-      <div class="right">
-        <p>PhisMeNot All Right reserved</p>
-        <p>© 2023</p>
-      </div>
     </div>
+    <footer>
+      <p>PhisMeNot 2023 All Right reserved</p>
+    </footer>
   </body>
 </html>
